+++ changelog.md
@@ -4,6 +4,7 @@
 - Access component instance's name with `vm.$name` (rather than `vm.$options.name`)  
 - Access component's html content with `vm.$html` (rather than `vm.$el.outerHTML`)
 - `vm.$find` allows you to search a component's html for matching elements or component instances.
+- `vm.$contains` tests whether a component contains another component or a html element.
 
 ## 0.2.0
 - stubComponents flag automatically stubs all *known* components